+++ Project3/TextLib.h
@@ -22,7 +22,7 @@ namespace Text
 	namespace Choose
 	{
 		int Horizontal(std::vector<std::string> text, int starting_position, COORD starting_point, TextAlign text_align, bool clear_after, Window &main_window);
-		int Veritcal(std::vector<std::string> text, int starting_position, COORD starting_posint, int spacing, TextAlign text_align, bool clear_after, Window &main_window);
+		int Veritcal(std::vector<std::string> text, short starting_position, COORD starting_posint, short spacing, TextAlign text_align, bool clear_after, Window &main_window);
 		int Numeric(int max, COORD starting_point, bool zero_allowed, Window &main_window);
 	}
